CoordinatorDashboardController

See merge request miw-se-cursisten/cohort-apr25/quizmaster/team-1/quizmaster!57

+++ src/main/java/controller/ManageCoursesController.java
@@ -3,12 +3,14 @@ package controller;
 import database.mysql.CourseDAO;
 import javafx.fxml.FXML;
 import javafx.scene.control.Alert;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.ListView;
 import model.Course;
 import model.User;
 import view.Main;
 
 import java.util.List;
+import java.util.Optional;
 
 public class ManageCoursesController {
 
@@ -56,15 +58,24 @@ public class ManageCoursesController {
             alert.setTitle("Fout bij verwijderen");
             alert.setHeaderText("Verwijderen mislukt");
             alert.setContentText("Selecteer een cursus om te verwijderen.");
-            alert.showAndWait();
-            return;
+            alert.show();
+        }else{
+            Alert comfirmAlert = new Alert(Alert.AlertType.CONFIRMATION);
+            comfirmAlert.setTitle("Bevestiging");
+            comfirmAlert.setHeaderText("Wil jij course "+ selectedCourse.getCourseName() +" verwijderen?");
+            Optional<ButtonType> result = comfirmAlert.showAndWait();
+
+            if(result.isPresent() && result.get() == ButtonType.OK){
+                courseDAO.deleteCourse(selectedCourse.getCourseName());
+                courseList.getItems().remove(selectedCourse);
+
+                Alert alert = new Alert(Alert.AlertType.INFORMATION);
+                alert.setTitle("Bevestiging");
+                alert.setHeaderText(null);
+                alert.setContentText("Course "+ selectedCourse.getCourseName() + " is verwijderd");
+                alert.show();
+            }
         }
-        courseDAO.deleteCourse(selectedCourse.getCourseName());
-        courseList.getItems().remove(selectedCourse);
-        Alert alert = new Alert(Alert.AlertType.INFORMATION);
-        alert.setTitle("Bevestiging");
-        alert.setHeaderText(null);
-        alert.setContentText("Course is verwijderd");
-        alert.showAndWait();
+
     }
 }
